Found bug in ReturnsAroundEarnings.fsx
For some reason, we can't seem to fetch returns for recent transcripts (failwith pops up for more recent calls). I think we had this sort of problem before, but since we had enough return observations at the time I just carried on to the classification script.

Still working on the comments